Allow capsules to be loaded from memory (#21)
Capsules now no longer need to exist on the file system in order to be loaded. Capsules now use the prefix `capsule_` followed by a hash of their contents as their module name, instead of their file name. This has the added benefit of making it possible to load different revisions of the same capsule at the same time.